Try to implement a TUI model using Termion... and fail.
I thought it'd be pretty straight forward to do this by managing the TUI
elements myself (e.g., window viewports, lists and selections) but it's
inherently tricky.

When investigating the rs-tui package initially I thought it was
overkill for such a simple TUI as what I need but now I'm thinking this
TUI isn't so simple (as one probably doesn't really exist) and so using
the rs-tui crate is the way to go.