Add context support to the highlighting.
Contexts mark sections of the mail body in which different regexes can
be grouped together.

Currently still clumsily declared in code.

The Highlighter now carries the current context state as it highlights,
requiring it to be passed around as a mutable reference which mightn't
be the best.  Perhaps the mutable context state and the actual
highlighting declarations can be separated further.